UPDATE: split logic in scrapper.go file
