--- scrapper/scrapper.js
@@ -1,2 +1,0 @@
-package scrapper
-
